+++ graphene_django_crud/input_types.py
@@ -3,7 +3,14 @@ import graphene
 from collections import OrderedDict
 
 
-def gen_inputType_filter(name, field, with_number_filter=False, with_text_filter=False):
+def gen_inputType_filter(
+    name,
+    field,
+    with_number_filter=False,
+    with_text_filter=False,
+    with_date_filter=False,
+    with_time_filter=False,
+):
 
     items = OrderedDict()
 
@@ -22,6 +29,17 @@ def gen_inputType_filter(name, field, with_number_filter=False, with_text_filter
         items["startswith"] = graphene.Field(field)
         items["endswith"] = graphene.Field(field)
         items["regex"] = graphene.String()
+
+    if with_date_filter:
+        items["year"] = graphene.Field(IntFilter)
+        items["month"] = graphene.Field(IntFilter)
+        items["day"] = graphene.Field(IntFilter)
+        items["week_day"] = graphene.Field(IntFilter)
+
+    if with_time_filter:
+        items["hour"] = graphene.Field(IntFilter)
+        items["minute"] = graphene.Field(IntFilter)
+        items["second"] = graphene.Field(IntFilter)
 
     return type(
         name,
@@ -45,11 +63,24 @@ StringFilter = gen_inputType_filter(
 )
 
 DateTimeFilter = gen_inputType_filter(
-    "DatetimeFilter", graphene.DateTime, with_number_filter=True, with_text_filter=False
+    "DatetimeFilter",
+    graphene.DateTime,
+    with_number_filter=True,
+    with_text_filter=False,
+    with_date_filter=True,
+    with_time_filter=True,
 )
 TimeFilter = gen_inputType_filter(
-    "timeFilter", graphene.Time, with_number_filter=True, with_text_filter=False
+    "timeFilter",
+    graphene.Time,
+    with_number_filter=True,
+    with_text_filter=False,
+    with_time_filter=True,
 )
 DateFilter = gen_inputType_filter(
-    "DateFilter", graphene.Date, with_number_filter=True, with_text_filter=False
+    "DateFilter",
+    graphene.Date,
+    with_number_filter=True,
+    with_text_filter=False,
+    with_date_filter=True,
 )